refactor(web): move dataset export into dataset_export pkg
export_core.py + shim; tests patch export_core globals. Refs #344 phase C.

Made-with: Cursor
