+++ app/web/services/dataset_export/__init__.py
@@ -1,0 +1,4 @@
+"""Bounded context: dataset export / retro-export / ZIP (#344).
+
+Реализация в ``export_core.py``; корневой shim — ``services/dataset_export_service.py``.
+"""
